refactor: move useFileUpload to customHooks folder
